% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eqtl_plotly.R
\name{eqtl_plotly}
\alias{eqtl_plotly}
\title{Locus eQTL scatter plotly}
\usage{
eqtl_plotly(loc, gene_filter = NULL, tissue_filter = NULL, height = NULL)
}
\arguments{
\item{loc}{Object of class 'locus' to use for plot. See \link{locus}.}

\item{gene_filter}{Character vector of genes to filter eQTL results.}

\item{tissue_filter}{Character vector of tissues to filter eQTL results.}

\item{height}{Controls the plot height.}
}
\value{
A \code{plotly} scatter plot.
}
\description{
Produces a scatter plot using plotly of embedded eQTL data acquired through
the LDlink API via \code{\link[=link_eqtl]{link_eqtl()}}.
}
\seealso{
\code{\link[=link_eqtl]{link_eqtl()}} \code{\link[=locus_plotly]{locus_plotly()}}
}
