% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eQTLplot.R
\name{eQTLplot}
\alias{eQTLplot}
\title{eQTL plot}
\usage{
eQTLplot(
  x,
  ...,
  filter_gene_name = NULL,
  filter_gene_biotype = NULL,
  pcutoff = 5e-08,
  col = "black",
  outline_col = NA,
  up_palette = "OrRd",
  down_palette = "Blues 3",
  alpha = 0.5,
  tissue = "Whole Blood",
  eqtl_gene = x$gene,
  xlab = NULL,
  ylab = expression("-log"[10] ~ "P"),
  cex.axis = 0.8,
  cex.text = 0.7,
  use_layout = TRUE,
  heights = c(3, 2),
  maxrows = 7,
  xticks = "bottom",
  border = FALSE,
  gene_col = "blue4",
  exon_col = "blue4",
  exon_border = "blue4",
  text_pos = "top",
  legend_pos = "topleft"
)
}
\arguments{
\item{x}{Object of class 'locus' to use for plot. See \link{locus}.}

\item{...}{Other arguments passed to \code{\link[=plot]{plot()}} for the scatter plot.}

\item{filter_gene_name}{Vector of gene names to display.}

\item{filter_gene_biotype}{Vector of gene biotypes to be filtered. Use
\code{\link[ensembldb:EnsDb-class]{ensembldb::listGenebiotypes()}} to display possible biotypes. For example,
\code{ensembldb::listGenebiotypes(EnsDb.Hsapiens.v75)}}

\item{pcutoff}{Cut-off for p value significance. Defaults to p = 5e-08. Set
to \code{NULL} to disable.}

\item{col}{Colour of points for SNPs which do not have eQTLs.}

\item{outline_col}{Colour of symbol outlines. Default is \code{NA} for no outlines.}

\item{up_palette}{Character string specifying palette for upregulation eQTL
using \link[grDevices:palettes]{grDevices::hcl.colors}}

\item{down_palette}{Character string specifying palette for downregulation
eQTL using \link[grDevices:palettes]{grDevices::hcl.colors}}

\item{alpha}{Alpha opacity for points}

\item{tissue}{GTex tissue in which eQTL has been measured}

\item{eqtl_gene}{Gene showing eQTL effect}

\item{xlab}{x axis title}

\item{ylab}{y axis title}

\item{cex.axis}{Specifies font size for axis numbering.}

\item{cex.text}{Font size for gene text.}

\item{use_layout}{Logical whether \code{graphics::layout} is called. Default
\code{TRUE} is for a standard single plot. Set to \code{FALSE} if a more complex
layout with multiple plots is required.}

\item{heights}{Ratio of top to bottom plot. See \link{layout}.}

\item{maxrows}{Specifies maximum number of rows to display in gene
annotation panel.}

\item{xticks}{Character value of either 'top' or 'bottom' specifying
whether x axis ticks and numbers are plotted on top or bottom plot window.}

\item{border}{Logical whether a bounding box is plotted around upper and
lower plots.}

\item{gene_col}{Colour for genes and exons.}

\item{exon_col}{Fill colour for exons.}

\item{exon_border}{Border line colour outlining exons. Set to \code{NA} for no
border.}

\item{text_pos}{Character value of either 'top' or 'left' specifying
placement of gene name labels.}

\item{legend_pos}{Position of legend. See \code{\link[=legend]{legend()}}. Set to \code{NULL} to hide
legend.}
}
\value{
No return value. Produces a plot using base graphics.
}
\description{
Experimental plotting function for overlaying eQTL data from GTex on top of
GWAS results. y axis shows the -log10 p-value for the GWAS result.
Significant eQTL for the specified gene are overlaid using colours and
symbols.
}
