% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locusplot.R
\name{genetracks}
\alias{genetracks}
\title{Plot gene tracks}
\usage{
genetracks(
  locus,
  border = FALSE,
  cex.axis = 0.8,
  cex.text = 0.7,
  genecol = "blue4",
  maxrows = 7,
  xticks = TRUE,
  xlab = NULL
)
}
\arguments{
\item{locus}{Object of class 'locus' generated by \code{\link[=locus]{locus()}}.}

\item{border}{Logical whether a bounding box is plotted.}

\item{cex.axis}{Specifies font size for axis numbering.}

\item{cex.text}{Font size for gene text.}

\item{genecol}{Colour for genes and exons.}

\item{maxrows}{Specifies maximum nunber of rows to display in gene
annotation panel.}

\item{xticks}{Logical whether x axis ticks and numbers are plotted.}

\item{xlab}{Title for x axis. Defaults to chromosome \code{seqname} specified
in \code{locus}.}
}
\value{
No return value.
}
\description{
Plot gene annotation tracks from ensembldb data.
}
\details{
This function is called by plot.locus(). It can used to plot the
gene annotation tracks on their own. It uses base graphics, so layout() can
be used to position adjacent plots above or below.
}
