% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_recomb.R
\name{get_recomb}
\alias{get_recomb}
\title{Query UCSC for Recombination data}
\usage{
get_recomb(loc, genome = "hg38", table = NULL)
}
\arguments{
\item{loc}{Object of class 'locus' generated by \code{\link[=locus]{locus()}}}

\item{genome}{Either \code{"hg38"} or \code{"hg19"}}

\item{table}{Optional character value specifying which recombination table to
use.}
}
\value{
A list object of class 'locus'. Recombination data is added as list
element \code{recomb}.
}
\description{
Adds recombination data by querying UCSC genome browser.
}
