% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gg_addgenes.R
\name{gg_addgenes}
\alias{gg_addgenes}
\title{Add gene tracks to a ggplot2 plot}
\usage{
gg_addgenes(p, loc, height = unit(5, "cm"), ...)
}
\arguments{
\item{p}{ggplot2 plot object}

\item{loc}{Object of class 'locus' to use for plot. See \code{\link[=locus]{locus()}}.}

\item{height}{A unit object specifying height of the lower gene track.}

\item{...}{Additional arguments passed to \code{\link[=gg_genetracks]{gg_genetracks()}} to control
colours of gene tracks etc.}
}
\value{
A gtable plotting object.
}
\description{
Adds gene tracks to an existing ggplot2 plot.
}
\details{
The combined plot is outputted to the current device. A gtable plotting object is
returned invisibly. This can be plotted using \code{\link[=grid.draw]{grid.draw()}}.
}
\seealso{
\code{\link[=gg_scatter]{gg_scatter()}} \code{\link[=gg_genetracks]{gg_genetracks()}}
}
