% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_ggplot.R
\name{locus_ggplot}
\alias{locus_ggplot}
\title{Locus plot using ggplot2}
\usage{
locus_ggplot(loc, height = unit(5, "cm"), xticks = "top")
}
\arguments{
\item{loc}{Object of class 'locus' to use for plot. See \code{\link[=locus]{locus()}}.}

\item{height}{A unit object specifying height of the lower gene track.}

\item{xticks}{Character value of either 'top' or 'bottom' specifying whether
x axis ticks and numbers are plotted on top or bottom plot window.}
}
\description{
Genomic locus plot similar to locuszoom.
}
\seealso{
\code{\link[=gg_scatter]{gg_scatter()}} \code{\link[=gg_genetracks]{gg_genetracks()}}
}
