% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locusplot.R
\name{locusplot}
\alias{locusplot}
\title{Locus plot}
\usage{
locusplot(
  data,
  xrange = NULL,
  seqname = NULL,
  gene = NULL,
  flank = 50000,
  ens_version = "EnsDb.Hsapiens.v75",
  chrom = "chrom",
  pos = "pos",
  p = "p",
  labs = "rsid",
  pcutoff = 5e-08,
  chromCols = "royalblue",
  sigCol = "red",
  xlab = NULL,
  ylab = "-log10 P value",
  cex.axis = 0.8,
  heights = c(3, 2),
  maxrows = 7,
  xticks = "bottom",
  index_snp = NULL,
  LD = TRUE,
  pop = "CEU",
  r2d = "r2",
  LDtoken = "",
  border = FALSE,
  LDcols = c("grey", "blue", "cyan", "green3", "orange", "red", "purple"),
  ...
)
}
\arguments{
\item{data}{Dataset (data.frame or data.table) to use for plot.}

\item{xrange}{Vector of genomic position range for the x axis.}

\item{seqname}{Specifies which chromosome to plot.}

\item{gene}{Optional specifies which gene to view. Either \code{xrange} with
\code{seqname}, or \code{gene} must be specified.}

\item{flank}{Vector of how much flanking region left and right of the gene
to show.}

\item{ens_version}{Specifies which ensembl database to query for gene and
exon positions. See \code{ensembldb} Bioconductor package.}

\item{chrom}{Determines which column in \code{data} contains chromosome
information.}

\item{pos}{Determines which column in \code{data} contains position information.}

\item{p}{Determines which column in \code{data} contains SNP p values.}

\item{labs}{Determines which column in \code{data} contains SNP rs IDs.}

\item{pcutoff}{Cut-off for p value significance. Defaults to 5e-08.}

\item{chromCols}{Colour for normal points if \code{LD} is \code{FALSE}.}

\item{sigCol}{Colour for significant points if \code{LD} is \code{FALSE}.}

\item{xlab}{x axis title.}

\item{ylab}{y axis title.}

\item{cex.axis}{Specifies font size for axis numbering.}

\item{heights}{Ratio of top to bottom plot. See \link{layout}.}

\item{maxrows}{Specifies maximum nunber of rows to display in gene
annotation panel.}

\item{xticks}{Character value specifying whether x axis ticks and numbers.}

\item{index_snp}{Specifies the index SNP for displaying linkage
disequilibrium (LD). If not specifiied, the SNP with the lowest P value is
selected.}

\item{LD}{Logical whether LD is plotted. Queries 1000 genomes via LDlinkR
package. See \link{LDlinkR}.}

\item{pop}{A 1000 Genomes Project population, (e.g. YRI or CEU), multiple
allowed, default = "CEU". Passed to \link[LDlinkR:LDmatrix]{LDlinkR::LDmatrix}.}

\item{r2d}{Either "r2" for LD r^2 or "d" for LD D', default = "r2". Passed
to \link[LDlinkR:LDmatrix]{LDlinkR::LDmatrix}.}

\item{LDtoken}{Peronsal access token for accessing 1000 genomes LD data via
LDlink API. See \link{LDlinkR}.}

\item{border}{Logical whether a bounding box is plotted around upper and
lower plots.}

\item{LDcols}{Vector of colours for plotting LD. The first colour is for SNPs
which lack LD information. The next 5 colours are for r2 or D' LD results
ranging from 0 to 1 in intervals of 0.2. The final colour is for the index
SNP.}

\item{...}{Other arguments passed to \code{\link[=plot]{plot()}} for the scatter plot.}
}
\value{
Returns a list containing the subset of data plotted, chromosome and
genomic position range.
}
\description{
Genomic locus plot similar to locuszoom.
}
\details{
This is an R version of locuszoom for generating publication ready Manhattan
plots of gene loci. It references ensembl databases for annotating genes
and exons. It queries LDlink to retrieve LD information on the index SNP.
}
