% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_layout.R
\name{multi_layout}
\alias{multi_layout}
\title{Layout multiple locus plots}
\usage{
multi_layout(nrow = 1, ncol = 1, heights = c(3, 2))
}
\arguments{
\item{nrow}{Number of rows of plots}

\item{ncol}{Number of columns of plots}

\item{heights}{Vector of length 2 specifying height for plot and gene tracks}
}
\value{
Sets \code{\link[=layout]{layout()}} to enable multiple plots per page. Returns \code{par()} so
that layout can be reset to default at the end of plotting.
}
\description{
Uses \code{\link[=layout]{layout()}} to set up a pages of multiple locus plots.
}
