% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_plot2.R
\name{overlay_plot2}
\alias{overlay_plot2}
\title{Plot overlaying eQTL and GWAS data}
\usage{
overlay_plot2(
  x,
  base_col = "black",
  up_palette = "OrRd",
  down_palette = "Blues 3",
  alpha = 0.5,
  tissue = "Whole Blood",
  eqtl_gene = x$gene,
  ...
)
}
\arguments{
\item{x}{Object of class 'locus' to use for plot. See \link{locus}.}

\item{base_col}{Colour of points for SNPs which do not have eQTLs.}

\item{up_palette}{Character string specifying palette for upregulation eQTL
using \link[grDevices:palettes]{grDevices::hcl.colors}}

\item{down_palette}{Character string specifying palette for downregulation
eQTL using \link[grDevices:palettes]{grDevices::hcl.colors}}

\item{alpha}{Alpha opacity for points}

\item{tissue}{GTex tissue in which eQTL has been measured}

\item{eqtl_gene}{Gene showing eQTL effect}

\item{...}{Other arguments passed to \code{\link[=locus_plot]{locus_plot()}} for the locus plot.}
}
\value{
No return value. Produces a plot using base graphics.
}
\description{
Experimental plotting function for overlaying eQTL data from GTEx on top of
GWAS results. y axis shows the -log10 p-value for the GWAS result.
Significant eQTL for the specified gene are overlaid using colours and
symbols.
}
