% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locusplot.R
\name{plot.locus}
\alias{plot.locus}
\title{Locus plot}
\usage{
\method{plot}{locus}(
  x,
  ...,
  filter_gene_name = NULL,
  filter_gene_biotype = NULL,
  pcutoff = 5e-08,
  chromCols = "royalblue",
  sigCol = "red",
  xlab = NULL,
  ylab = expression("-log"[10] ~ "P"),
  cex.axis = 0.8,
  cex.text = 0.7,
  heights = c(3, 2),
  maxrows = 7,
  xticks = "bottom",
  border = FALSE,
  LDcols = c("grey", "royalblue", "cyan2", "green3", "orange", "red", "purple"),
  gene_col = "blue4",
  legend_pos = "topleft"
)
}
\arguments{
\item{x}{Object of class 'locus' to use for plot. See \link{locus}.}

\item{...}{Other arguments passed to \code{\link[=plot]{plot()}} for the scatter plot.}

\item{filter_gene_name}{Vector of gene names to display.}

\item{filter_gene_biotype}{Vector of gene biotypes to be filtered. Use
\code{\link[ensembldb:EnsDb-class]{ensembldb::listGenebiotypes()}} to display possible biotypes. For example,
\code{ensembldb::listGenebiotypes(EnsDb.Hsapiens.v75)}}

\item{pcutoff}{Cut-off for p value significance. Defaults to p = 5e-08. Set
to \code{NULL} to disable.}

\item{chromCols}{Colour for normal points if \code{LD} is \code{FALSE} when the locus
object is made.}

\item{sigCol}{Colour for significant points if \code{LD} is \code{FALSE}.}

\item{xlab}{x axis title.}

\item{ylab}{y axis title.}

\item{cex.axis}{Specifies font size for axis numbering.}

\item{cex.text}{Font size for gene text.}

\item{heights}{Ratio of top to bottom plot. See \link{layout}.}

\item{maxrows}{Specifies maximum nunber of rows to display in gene
annotation panel.}

\item{xticks}{Character value of either 'top' or 'bottom' specifying
whether x axis ticks and numbers are plotted on top or bottom plot window.}

\item{border}{Logical whether a bounding box is plotted around upper and
lower plots.}

\item{LDcols}{Vector of colours for plotting LD. The first colour is for SNPs
which lack LD information. The next 5 colours are for r2 or D' LD results
ranging from 0 to 1 in intervals of 0.2. The final colour is for the index
SNP.}

\item{gene_col}{Colour for genes and exons.}

\item{legend_pos}{Position of legend. See \code{\link[=legend]{legend()}}. Set to \code{NULL} to hide
legend.}

\item{exon_col}{Fill colour for exons.}

\item{exon_border}{Border line colour outlining exons. Set to \code{NA} for no
border.}
}
\value{
No return value.
}
\description{
Genomic locus plot similar to locuszoom.
}
\details{
This is an R version of locuszoom for generating publication ready Manhattan
plots of gene loci. It references ensembl databases for annotating genes
and exons. It queries LDlink to retrieve LD information on the index SNP.
Use \code{\link[=locus]{locus()}} first to generate an object of class 'locus' for plotting.
}
