% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanity_check.R
\name{sanity_check}
\alias{sanity_check}
\title{Sanity checks on GWAS dataframes}
\usage{
sanity_check(data, data2, chrom = NULL, pos = NULL, p = NULL, labs = NULL)
}
\arguments{
\item{data}{Dataframe of GWAS results with columns for chromosome, position,
p value and SNP rs IDs. Data.tables are coerced to dataframe.}

\item{data2}{Dataframe of 2nd GWAS results for comparison.}

\item{chrom}{Vector determines which column in \code{data} and \code{data2} contains
chromosome information. If \code{NULL} or \code{NA} tries to autodetect the columns.
The 1st element refers to \code{data} and 2nd element refers to \code{data2}.}

\item{pos}{Determines which column in \code{data} and \code{data2} contains position
information. See \code{chrom}.}

\item{p}{Determines which column in \code{data} and \code{data2} contains SNP
p-values. See \code{chrom}.}

\item{labs}{Determines which column in \code{data} and \code{data2} contains SNP rs
IDs. See \code{chrom}.}
}
\value{
No return value. Prints diagnostic messages.
}
\description{
Performs sanity checks on GWAS dataframes to check similar chromosome
content, overlap in SNP ids and alignment of SNPs by chromosome/position.
}
