% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_plot.R
\name{scatter_plot}
\alias{scatter_plot}
\title{Locus scatter plot}
\usage{
scatter_plot(
  x,
  pcutoff = 5e-08,
  chromCol = "royalblue",
  sigCol = "red",
  xlab = NULL,
  ylab = expression("-log"[10] ~ "P"),
  cex.axis = 1,
  xticks = FALSE,
  border = FALSE,
  showLD = TRUE,
  LD_scheme = c("grey", "royalblue", "cyan2", "green3", "orange", "red", "purple"),
  legend_pos = "topleft",
  add = FALSE,
  align = TRUE,
  ...
)
}
\arguments{
\item{x}{Object of class 'locus' to use for plot. See \link{locus}.}

\item{pcutoff}{Cut-off for p value significance. Defaults to p = 5e-08. Set
to \code{NULL} to disable.}

\item{chromCol}{Colour for normal points if \code{LD} is \code{FALSE} when the locus
object is made.}

\item{sigCol}{Colour for significant points if \code{LD} is \code{FALSE}.}

\item{xlab}{x axis title.}

\item{ylab}{y axis title.}

\item{cex.axis}{Specifies font size for axis numbering.}

\item{xticks}{Logical whether x axis numbers and axis title are plotted.}

\item{border}{Logical whether a bounding box is plotted around upper and
lower plots.}

\item{showLD}{Logical whether to show LD with colours}

\item{LD_scheme}{Vector of colours for plotting LD. The first colour is for SNPs
which lack LD information. The next 5 colours are for r2 or D' LD results
ranging from 0 to 1 in intervals of 0.2. The final colour is for the index
SNP.}

\item{legend_pos}{Position of legend. See \code{\link[=legend]{legend()}}. Set to \code{NULL} to hide
legend.}

\item{add}{Logical whether to add points to an existing plot or generate a
new plot.}

\item{align}{Logical whether to set \code{\link[=par]{par()}} to align the plot.}

\item{...}{Other arguments passed to \code{\link[=plot]{plot()}} for the scatter plot.}
}
\value{
No return value. Produces a scatter plot using base graphics.
}
\description{
Produces a scatter plot from a 'locus' class object. Intended for use with
\code{\link[=set_layers]{set_layers()}}.
}
\seealso{
\code{\link[=locus]{locus()}} \code{\link[=set_layers]{set_layers()}}
}
