% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SLE_gwas_sub.R
\docType{data}
\name{SLE_gwas_sub}
\alias{SLE_gwas_sub}
\title{SLE GWAS data subset}
\format{
Data frame with 1990 rows and 11 variables
}
\source{
\url{https://www.ebi.ac.uk/gwas/studies/GCST003156}
}
\usage{
data(SLE_gwas_sub)
}
\description{
Dataset of SNPs at 3 gene loci (UBE2L3, STAT4, IRF5) from GWAS on SLE
(Bentham et al, 2015, Nature Genetics 47(12):1457-64, PMID: 26502338).
}
\keyword{datasets}
