% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locusplot.R
\name{genetracks}
\alias{genetracks}
\title{Plot gene tracks}
\usage{
genetracks(
  locus,
  filter_gene_name = NULL,
  filter_gene_biotype = NULL,
  border = FALSE,
  cex.axis = 0.8,
  cex.text = 0.7,
  gene_col = "blue4",
  exon_col = "blue4",
  exon_border = "blue4",
  maxrows = NULL,
  xticks = TRUE,
  xlab = NULL
)
}
\arguments{
\item{locus}{Object of class 'locus' generated by \code{\link[=locus]{locus()}}.}

\item{filter_gene_name}{Vector of gene names to display.}

\item{filter_gene_biotype}{Vector of gene biotypes to be filtered. Use
\code{\link[ensembldb:EnsDb-class]{ensembldb::listGenebiotypes()}} to display possible biotypes. For example,
\code{ensembldb::listGenebiotypes(EnsDb.Hsapiens.v75)}}

\item{border}{Logical whether a bounding box is plotted.}

\item{cex.axis}{Specifies font size for axis numbering.}

\item{cex.text}{Font size for gene text.}

\item{gene_col}{Colour for gene lines.}

\item{exon_col}{Fill colour for exons.}

\item{exon_border}{Border line colour outlining exons. Set to \code{NA} for no
border.}

\item{maxrows}{Specifies maximum nunber of rows to display in gene
annotation panel.}

\item{xticks}{Logical whether x axis ticks and numbers are plotted.}

\item{xlab}{Title for x axis. Defaults to chromosome \code{seqname} specified
in \code{locus}.}
}
\value{
No return value.
}
\description{
Plot gene annotation tracks from ensembldb data.
}
\details{
This function is called by plot.locus(). It can used to plot the
gene annotation tracks on their own. It uses base graphics, so layout() can
be used to position adjacent plots above or below.
}
