% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus.R
\name{locus}
\alias{locus}
\title{Create locus object for plotting}
\usage{
locus(
  data,
  xrange = NULL,
  seqname = NULL,
  gene = NULL,
  flank = 50000,
  ens_version = "EnsDb.Hsapiens.v75",
  chrom = "chrom",
  pos = "pos",
  p = "p",
  labs = "rsid",
  index_snp = NULL,
  LD = TRUE,
  eQTL = FALSE,
  pop = "CEU",
  r2d = "r2",
  LDtoken = ""
)
}
\arguments{
\item{data}{Dataset (data.frame or data.table) to use for plot.}

\item{xrange}{Vector of genomic position range for the x axis.}

\item{seqname}{Specifies which chromosome to plot.}

\item{gene}{Optional specifies which gene to view. Either \code{xrange} with
\code{seqname}, or \code{gene} must be specified.}

\item{flank}{Single value or vector with 2 values for how much flanking
region left and right of the gene to show.}

\item{ens_version}{Specifies which Ensembl database to query for gene and
exon positions. See \code{ensembldb} Bioconductor package.}

\item{chrom}{Determines which column in \code{data} contains chromosome
information. Automatically looks for PLINK headings.}

\item{pos}{Determines which column in \code{data} contains position information.
Automatically looks for PLINK headings.}

\item{p}{Determines which column in \code{data} contains SNP p values.
Automatically looks for PLINK headings.}

\item{labs}{Determines which column in \code{data} contains SNP rs IDs.
Automatically looks for PLINK headings.}

\item{index_snp}{Specifies the index SNP for displaying linkage
disequilibrium (LD). If not specified, the SNP with the lowest P value is
selected.}

\item{LD}{Logical whether LD is plotted. Queries 1000 Genomes via LDlinkR
package. See \link[LDlinkR:LDmatrix]{LDlinkR::LDmatrix}. Results are cached using the \code{memoise}
package, so that if exactly the same locus is requested the system does not
repeatedly call the API.}

\item{eQTL}{Logical whether to obtain eQTL data. Queries GTEx eQTL data via
\link[LDlinkR:LDexpress]{LDlinkR::LDexpress} using the SNP specified by \code{index_snp}}

\item{pop}{A 1000 Genomes Project population, (e.g. YRI or CEU), multiple
allowed, default = "CEU". Passed to \link[LDlinkR:LDmatrix]{LDlinkR::LDmatrix}.}

\item{r2d}{Either "r2" for LD r^2 or "d" for LD D', default = "r2". Passed
to \link[LDlinkR:LDmatrix]{LDlinkR::LDmatrix}.}

\item{LDtoken}{Personal access token for accessing 1000 Genomes LD data via
LDlink API. See \link{LDlinkR}.}
}
\value{
Returns an object of class 'locus' ready for plotting, containing
the subset of GWAS data to be plotted,
chromosome and genomic position range,
Ensembl database version number,
column names for chromosome, position, SNP ID, p value
locus gene information from Ensembl and
locus exon information from Ensembl.
}
\description{
Creates object of class 'locus' for genomic locus plot similar to
\code{locuszoom}.
}
\details{
This is an R version of \code{locuszoom} (http://locuszoom.org) for generating
publication ready Manhattan plots of gene loci. It references Ensembl
databases using the \code{ensembldb} Bioconductor package framework for annotating
genes and exons. It queries LDlink (https://ldlink.nci.nih.gov/) via the
\link{LDlinkR} package to retrieve linkage disequilibrium (LD) information on a
reference SNP.
}
\examples{
library(EnsDb.Hsapiens.v75)
data(SLE_gwas_sub)
loc <- locus(SLE_gwas_sub, gene = 'UBE2L3', flank = 1e5, LD = FALSE)
plot(loc)
loc2 <- locus(SLE_gwas_sub, gene = 'STAT4', flank = 1e5, LD = FALSE)
plot(loc2)
}
