% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locusplot.R
\name{locus}
\alias{locus}
\title{Create locus object for plotting}
\usage{
locus(
  data,
  xrange = NULL,
  seqname = NULL,
  gene = NULL,
  flank = 50000,
  ens_version = "EnsDb.Hsapiens.v75",
  chrom = "chrom",
  pos = "pos",
  p = "p",
  labs = "rsid",
  index_snp = NULL,
  LD = TRUE,
  pop = "CEU",
  r2d = "r2",
  LDtoken = ""
)
}
\arguments{
\item{data}{Dataset (data.frame or data.table) to use for plot.}

\item{xrange}{Vector of genomic position range for the x axis.}

\item{seqname}{Specifies which chromosome to plot.}

\item{gene}{Optional specifies which gene to view. Either \code{xrange} with
\code{seqname}, or \code{gene} must be specified.}

\item{flank}{Vector of how much flanking region left and right of the gene
to show.}

\item{ens_version}{Specifies which ensembl database to query for gene and
exon positions. See \code{ensembldb} Bioconductor package.}

\item{chrom}{Determines which column in \code{data} contains chromosome
information. Automatically looks for PLINK headings.}

\item{pos}{Determines which column in \code{data} contains position information.
Automatically looks for PLINK headings.}

\item{p}{Determines which column in \code{data} contains SNP p values.
Automatically looks for PLINK headings.}

\item{labs}{Determines which column in \code{data} contains SNP rs IDs.
Automatically looks for PLINK headings.}

\item{index_snp}{Specifies the index SNP for displaying linkage
disequilibrium (LD). If not specifiied, the SNP with the lowest P value is
selected.}

\item{LD}{Logical whether LD is plotted. Queries 1000 genomes via LDlinkR
package. See \link{LDlinkR}. Results are cached using the \code{memoise} package, so
that if exactly the same locus is requested the system does not repeatedly
call the API.}

\item{pop}{A 1000 Genomes Project population, (e.g. YRI or CEU), multiple
allowed, default = "CEU". Passed to \link[LDlinkR:LDmatrix]{LDlinkR::LDmatrix}.}

\item{r2d}{Either "r2" for LD r^2 or "d" for LD D', default = "r2". Passed
to \link[LDlinkR:LDmatrix]{LDlinkR::LDmatrix}.}

\item{LDtoken}{Peronsal access token for accessing 1000 genomes LD data via
LDlink API. See \link{LDlinkR}.}
}
\value{
Returns an object of class 'locus' ready for plotting, containing
the subset of GWAS data to be plotted,
chromosome and genomic position range,
ensembl database version number,
column names for chromosome, position, SNP ID, p value
locus gene information from ensembl and
locus exon information from ensembl.
}
\description{
Creates object of class 'locus' for genomic locus plot similar to locuszoom.
}
\details{
This is an R version of locuszoom for generating publication ready Manhattan
plots of gene loci. It references ensembl databases using the ensembldb
Bioconductor package framework for annotating genes and exons. It queries
LDlink (https://ldlink.nci.nih.gov/) via the \link{LDlinkR} package to retrieve
LD information on a reference SNP.
}
