% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locus_plotly.R
\name{locus_plotly}
\alias{locus_plotly}
\title{Locus plotly}
\usage{
locus_plotly(loc, heights = c(0.6, 0.4), ...)
}
\arguments{
\item{loc}{Object of class 'locus' to use for plot. See \code{\link[=locus]{locus()}}.}

\item{heights}{Vector controlling relative height of each panel on 0-1 scale.}

\item{...}{Optional arguments passed to \code{\link[=genetrack_ly]{genetrack_ly()}} to control the gene
tracks.}
}
\description{
Genomic locus plot similar to locuszoom, using plotly.
}
\details{
This is an R/plotly version of locuszoom for generating publication ready
Manhattan plots of gene loci. It references Ensembl databases for annotating
genes and exons. Use \code{\link[=locus]{locus()}} first to generate an object of class 'locus'
for plotting. LDlink web server can be queried using function \code{\link[=link_LD]{link_LD()}} to
retrieve linkage disequilibrium (LD) information on the index SNP.
}
