% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/overlay_plotly.R
\name{overlay_plotly}
\alias{overlay_plotly}
\title{Plot overlaying eQTL and GWAS data using plotly}
\usage{
overlay_plotly(
  loc,
  gene_filter = NULL,
  tissue_filter = NULL,
  pcutoff = 5e-08,
  scheme = c("grey", "dodgerblue", "red"),
  xlab = NULL,
  ylab = NULL,
  marker_outline = "grey",
  marker_size = 7,
  recomb_col = "blue",
  showlegend = TRUE,
  height = NULL,
  webGL = TRUE
)
}
\arguments{
\item{loc}{Object of class 'locus' to use for plot. See \link{locus}.}

\item{gene_filter}{Character vector of genes to filter eQTL results.}

\item{tissue_filter}{Character vector of tissues to filter eQTL results.}

\item{pcutoff}{Cut-off for p value significance. Defaults to p = 5e-08. Set
to \code{NULL} to disable.}

\item{scheme}{Vector of 3 colours if LD is not shown: 1st = normal points,
2nd = colour for significant points, 3rd = index SNP(s).}

\item{xlab}{x axis title.}

\item{ylab}{y axis title.}

\item{marker_outline}{Specifies colour for outlining points.}

\item{marker_size}{Value for size of markers in plotly units.}

\item{recomb_col}{Colour for recombination rate line if recombination rate
data is present. Set to \code{NA} to hide the line. See \code{\link[=link_recomb]{link_recomb()}} to add
recombination rate data.}

\item{showlegend}{Logical whether to show a legend for the scatter points.}

\item{height}{Height in pixels (optional, defaults to automatic sizing).}

\item{webGL}{Logical whether to use webGL or SVG for scatter plot.}
}
\value{
A \code{plotly} scatter plot.
}
\description{
Produces a scatter plot using plotly of embedded eQTL data acquired through
the LDlink API via \code{\link[=link_eqtl]{link_eqtl()}} overlaid on GWAS data. As each SNP may
have eQTLs with multiple genes in multiple tissues, the method used is to
select the gene/tissue eQTL with the lowest p-value. SNPs are matched by
rsID.
}
\seealso{
\code{\link[=link_eqtl]{link_eqtl()}} \code{\link[=locus_plotly]{locus_plotly()}}
}
