% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locusplot.R
\name{plot.locus}
\alias{plot.locus}
\title{Locus plot}
\usage{
\method{plot}{locus}(
  x,
  ...,
  pcutoff = 5e-08,
  chromCols = "royalblue",
  sigCol = "red",
  xlab = NULL,
  ylab = expression("-log"[10] ~ "P"),
  cex.axis = 0.8,
  cex.text = 0.7,
  heights = c(3, 2),
  maxrows = 7,
  xticks = "bottom",
  border = FALSE,
  LDcols = c("grey", "blue", "cyan2", "green3", "orange", "red", "purple"),
  genecol = "blue4"
)
}
\arguments{
\item{x}{Object of class 'locus' to use for plot. See \link{locus}.}

\item{...}{Other arguments passed to \code{\link[=plot]{plot()}} for the scatter plot.}

\item{pcutoff}{Cut-off for p value significance. Defaults to 5e-08.}

\item{chromCols}{Colour for normal points if \code{LD} is \code{FALSE} when the locus
object is made.}

\item{sigCol}{Colour for significant points if \code{LD} is \code{FALSE}.}

\item{xlab}{x axis title.}

\item{ylab}{y axis title.}

\item{cex.axis}{Specifies font size for axis numbering.}

\item{cex.text}{Font size for gene text.}

\item{heights}{Ratio of top to bottom plot. See \link{layout}.}

\item{maxrows}{Specifies maximum nunber of rows to display in gene
annotation panel.}

\item{xticks}{Character value of either 'top' or 'bottom' specifying
whether x axis ticks and numbers are plotted on top or bottom plot window.}

\item{border}{Logical whether a bounding box is plotted around upper and
lower plots.}

\item{LDcols}{Vector of colours for plotting LD. The first colour is for SNPs
which lack LD information. The next 5 colours are for r2 or D' LD results
ranging from 0 to 1 in intervals of 0.2. The final colour is for the index
SNP.}

\item{genecol}{Colour for genes and exons.}
}
\value{
No return value.
}
\description{
Genomic locus plot similar to locuszoom.
}
\details{
This is an R version of locuszoom for generating publication ready Manhattan
plots of gene loci. It references ensembl databases for annotating genes
and exons. It queries LDlink to retrieve LD information on the index SNP.
}
