% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.R
\name{zoom}
\alias{zoom}
\title{Zoom browser to explore GWAS/eQTL results}
\usage{
zoom(
  data,
  ens_db,
  chrom = NULL,
  pos = NULL,
  p = NULL,
  labs = NULL,
  data2 = NULL,
  traits = NULL,
  scheme = c("royalblue", "skyblue", "red"),
  scheme2 = c("#33a02c", "#b2df8a", "purple"),
  pcutoff = 5e-08,
  eqtl_gene = NULL,
  eqtl_beta = NULL,
  eqtl_scheme = c("#FF0000", "#00FFFF", "#FF9000", "#0080FF", "#FFFF00", "#0000FF",
    "#80DD00", "#8000FF", "#009900", "#FF00FF"),
  add_hover = NULL,
  mh_points = 1e+05,
  recomb = NULL,
  ld_token = Sys.getenv("LDLINK_TOKEN"),
  ld_pop = "EUR",
  seq_filter = c(1:22, "X", "Y"),
  AnnotationDb = "org.Hs.eg.db"
)
}
\arguments{
\item{data}{Dataframe of GWAS results with columns for chromosome, position,
p value and SNP rs IDs. Data.tables are coerced to dataframe.}

\item{ens_db}{Either a character string which specifies which Ensembl
database package (version 86 and earlier for Homo sapiens) to query for
gene and exon positions (see \code{ensembldb} Bioconductor package). Or an
\code{ensembldb} object which can be obtained from the AnnotationHub database.
See the vignette and the \code{AnnotationHub} Bioconductor package for how to
create this object.}

\item{chrom}{Determines which column in \code{data} contains chromosome
information. If \code{NULL} or \code{NA} tries to autodetect the column. If \code{data2}
is provided, this is a vector where 1st element refers to \code{data} and 2nd
element refers to \code{data2}.}

\item{pos}{Determines which column in \code{data} (and optionally \code{data2})
contains position information. See \code{chrom}.}

\item{p}{Determines which column in \code{data} (and optionally \code{data2}) contains
SNP p-values. See \code{chrom}.}

\item{labs}{Determines which column in \code{data} (and optionally \code{data2})
contains SNP rs IDs. See \code{chrom}.}

\item{data2}{Optional 2nd dataframe of GWAS results for comparison.}

\item{traits}{Vector of trait names for identifying \code{data} and \code{data2}
datasets.}

\item{scheme}{Vector of 3 colours for main Manhattan plot: 1st, 2nd colours
for alternating chromosomes, 3rd colour for significant points.}

\item{scheme2}{Vector of colours for optional 2nd Manhattan plot.}

\item{pcutoff}{Cut-off for p value significance. Defaults to p = 5e-08. Set
to \code{NULL} to disable.}

\item{eqtl_gene}{Determines which column in \code{data} contains eQTL genes.}

\item{eqtl_beta}{Optional column name for beta coefficient to display upward
triangles for positive beta and downward triangles for negative beta
(significant SNPs only).}

\item{eqtl_scheme}{Colour scheme for eQTL genes.}

\item{add_hover}{Optional vector of column names in 'data' to add to the
plotly hover text for scatter points.}

\item{mh_points}{Number of points to display in manhattan plot. Default is
\code{1e5}.}

\item{recomb}{Optional \code{GRanges} class object of recombination data.}

\item{ld_token}{Personal access token for the LDlink API, available from
\url{https://ldlink.nih.gov/?tab=apiaccess}. See \code{LDlinkR} package
documentation and \code{\link[=link_LD]{link_LD()}}. When empty the LD controls are hidden. LD
information can only be requested if \code{eqtl_gene} is left as \code{NULL}. LD is
fetched on demand, not automatically, by pressing the "Get LD" button.}

\item{ld_pop}{1000 Genomes population used for LD. Defaults to \code{"EUR"}. See
\code{LDlinkR::LDproxy()} for the available codes.}

\item{seq_filter}{Vector of acceptable chromosomes. Used to restrict queries
to standard chromosome assembly.}

\item{AnnotationDb}{An \code{AnnotationDb} gene annotation database, specified
either as a character string or as an \code{AnnotationDb} class object, used to
obtain expanded gene names. The ensembl database specified in \code{ens_db} is
queried first. Set to \code{NULL} to disable this feature.}
}
\value{
No return value. Opens an interactive shiny window.
}
\description{
Interactive genome browser to explore GWAS/eQTL results using a shiny
interface.
}
\details{
This launches a shiny app to explore the GWAS/eQTL results through visualising
the Manhattan plot and exploring regional Manhattan plots of gene loci
through selecting points or searching SNPs/genes.

Linkage disequilibrium data can be pulled from LDlink API within the app
using the 'Get LD' button in the Settings dropdown. These buttons only appear
once a user provides a LDlink API token via the \code{ld_token} argument. Each API
request takes around 5-10 secs. 'Get LD' pins the current index SNP as the
reference variant and colours points by r^2 with it. The reference stays
pinned while you pan and zoom, so the colouring keeps its meaning and repeat
queries are served from the \code{memoise} cache rather than the API.
}
