% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zoom.R
\name{zoom}
\alias{zoom}
\title{Zoom browser to explore GWAS/eQTL results}
\usage{
zoom(
  data,
  ens_db,
  chrom = NULL,
  pos = NULL,
  p = NULL,
  labs = NULL,
  scheme = c("royalblue", "skyblue", "red"),
  pcutoff = 5e-08,
  eqtl_gene = NULL,
  eqtl_beta = NULL,
  eqtl_scheme = c("#FF0000", "#00FFFF", "#FF9000", "#0080FF", "#FFFF00", "#0000FF",
    "#80DD00", "#8000FF", "#009900", "#FF00FF"),
  add_hover = NULL,
  mh_points = 1e+05,
  recomb = NULL,
  token = NULL,
  seq_filter = c(1:22, "X", "Y"),
  AnnotationDb = "org.Hs.eg.db",
  ...
)
}
\arguments{
\item{data}{Dataframe of GWAS results with columns for chromosome, position,
p value and SNP rs IDs. Data.tables are coerced to dataframe.}

\item{ens_db}{Either a character string which specifies which Ensembl
database package (version 86 and earlier for Homo sapiens) to query for
gene and exon positions (see \code{ensembldb} Bioconductor package). Or an
\code{ensembldb} object which can be obtained from the AnnotationHub database.
See the vignette and the \code{AnnotationHub} Bioconductor package for how to
create this object.}

\item{chrom}{Determines which column in \code{data} contains chromosome
information If \code{NULL} tries to autodetect the column.}

\item{pos}{Determines which column in \code{data} contains position information.
If \code{NULL} tries to autodetect the column.}

\item{p}{Determines which column in \code{data} contains SNP p-values. If \code{NULL}
tries to autodetect the column.}

\item{labs}{Determines which column in \code{data} contains SNP rs IDs. If \code{NULL}
tries to autodetect the column.}

\item{scheme}{Vector of 3 colours: 1st = normal points, 2nd = colour for
significant points, 3rd = index SNP(s).}

\item{pcutoff}{Cut-off for p value significance. Defaults to p = 5e-08. Set
to \code{NULL} to disable.}

\item{eqtl_gene}{Determines which column in \code{data} contains eQTL genes.}

\item{eqtl_beta}{Optional column name for beta coefficient to display upward
triangles for positive beta and downward triangles for negative beta
(significant SNPs only).}

\item{eqtl_scheme}{Colour scheme for eQTL genes.}

\item{add_hover}{Optional vector of column names in 'data' to add to the
plotly hover text for scatter points.}

\item{mh_points}{Number of points to display in manhattan plot. Default is
\code{1e5}.}

\item{recomb}{Optional \code{GRanges} class object of recombination data.}

\item{token}{Personal access token for accessing 1000 Genomes LD data via
LDlink API. See \code{LDlinkR} package documentation and \code{\link[=link_LD]{link_LD()}}. LD
information can only be requested if \code{eqtl_gene} is left as \code{NULL}.}

\item{seq_filter}{Vector of acceptable chromosomes. Used to restrict queries
to standard chromosome assembly.}

\item{AnnotationDb}{An \code{AnnotationDb} gene annotation database, specified
either as a character string or as an \code{AnnotationDb} class object, used to
obtain expanded gene names. The ensembl database specified in \code{ens_db} is
queried first. Set to \code{NULL} to disable this feature.}

\item{...}{Optional arguments such as \code{pop} passed to \code{\link[=link_LD]{link_LD()}}.}
}
\value{
No return value. Opens an interactive shiny window.
}
\description{
Interactive genome browser to explore GWAS/eQTL results using a shiny
interface.
}
\details{
This launches a shiny app to explore the GWAS/eQTL results through visualising
the Manhattan plot and exploring regional Manhattan plots of gene loci
through selecting points or searching SNPs/genes.

Linkage disequilibrium data can be pulled from LDlink API within the app
using the 'Get LD' button in the Settings dropdown. These buttons only appear
once a user provides a LDlink API token via the \code{token} argument. Each
API request takes around 5-10 secs.
}
